refactor(team-host): extract shared overlay-binary provisioning to @myco/host
Task 2.2 (member) needs the same Tailscale (data-plane) provisioning + fetcher/
runner seams that Task 2.1 (host) built, but the member's `myco join` runs on
the @myco binary and @myco cannot import @myco-team (that edge is a cycle — the
dependency runs @myco-team → @myco). So the shared, provider-agnostic pieces move
to @myco/host/overlay-binaries.ts (one source of truth); @myco-team/host/
binaries.ts keeps Headscale-only provisioning and re-exports the shared symbols
so its callers + tests are unchanged.
