+++ packages/myco-team/src/host/run.ts
@@ -1,36 +1,12 @@
 /**
  * The real {@link CommandRunner} for Team Host orchestration — a thin spawn
  * wrapper shared by binary provisioning, the system-service supervisor, and the
- * headscale/tailscale CLI seams. Mirrors `@myco/service/run-command`'s
- * combined-output decoding but adds optional stdin (`input`), which the headscale
- * preauth-key mint and secret-put style calls need.
+ * headscale/tailscale CLI seams.
  *
- * NEVER rejects on a non-zero exit — the exit code is returned so callers decide
- * what a failure means (an idempotent step tolerates "already exists"; a hard
- * step surfaces it). A spawn error (ENOENT) resolves as exit 127 so a missing
- * binary reads as a normal failure, not an unhandled rejection.
+ * The implementation now lives in the shared `@myco/host/overlay-binaries.ts`
+ * (as `realCommandRunner`) so the host and the member (`myco join`, Task 2.2)
+ * share ONE runner. Re-exported here under the name `overlay.ts` and the rest of
+ * the host orchestration already import.
  */
-import { spawn } from 'node:child_process';
-
-import type { CommandRunner } from './binaries.js';
-
-export const realRunner: CommandRunner = {
-  run(command: string, args: string[], opts?: { input?: string }): Promise<{ stdout: string; exitCode: number }> {
-    return new Promise((resolve) => {
-      const child = spawn(command, args, { stdio: [opts?.input !== undefined ? 'pipe' : 'ignore', 'pipe', 'pipe'] });
-      const out: Buffer[] = [];
-      const err: Buffer[] = [];
-      child.stdout?.on('data', (b: Buffer) => out.push(b));
-      child.stderr?.on('data', (b: Buffer) => err.push(b));
-      child.on('error', (e: Error) => resolve({ stdout: String(e.message), exitCode: 127 }));
-      child.on('close', (code) => resolve({
-        stdout: Buffer.concat(out).toString('utf8') + Buffer.concat(err).toString('utf8'),
-        exitCode: code ?? 0,
-      }));
-      if (opts?.input !== undefined && child.stdin) {
-        child.stdin.write(opts.input);
-        child.stdin.end();
-      }
-    });
-  },
-};
+export { realCommandRunner as realRunner } from '@myco/host/overlay-binaries.js';
+export type { CommandRunner } from '@myco/host/overlay-binaries.js';
